MAJOR REROUTING HERE
The old main editing page has moved to `/substance`.
The "about" page is now the home page at `/`.
