+++ src/views/Substance.vue
@@ -1,0 +1,23 @@
+<template>
+  <div class="substance">
+    <ChemicalEditors />
+  </div>
+</template>
+
+<script>
+import ChemicalEditors from "@/components/ChemicalEditors";
+
+export default {
+  name: "substance",
+  components: {
+    ChemicalEditors
+  }
+};
+</script>
+
+<style scoped>
+.substance {
+  width: 1024px;
+  margin: auto;
+}
+</style>
