+++ tests/e2e/specs/home.js
@@ -1,111 +1,27 @@
-describe("The home page", () => {
+describe("The home page as seen by a visitor", () => {
+  beforeEach(() => {
+    cy.visit("/");
+  });
+  it("should have lorem ipsum", () => {
+    cy.contains("p", "Lorem ipsum dolor sit amet");
+  });
+  it("should display the login form to a non-authenticated user", () => {
+    cy.get('#login-card')
+  });
+  it("should not display the search bar form to a non-authenticated user", () => {
+    cy.get('[data-cy=search-box]').should('not.exist');
+  });
+});
+
+describe("The home page as seen by an authenticated user", () => {
   beforeEach(() => {
     cy.adminLogin();
     cy.visit("/");
   });
-  it("should have welcome msg", () => {
-    cy.contains("h1", "Welcome to ChemReg");
+  it("should not display the login form to an authenticated user", () => {
+    cy.get('#login-card').should('not.exist');
   });
-  it("should have dropdown", () => {
-    cy.get("#compound-type-dropdown").contains("defined");
-    cy.get("#compound-type-dropdown").contains("ill-defined");
-  });
-  it("should toggle ketcher/marvinjs on dropdown", () => {
-    cy.get("#compound-type-dropdown").select("defined");
-    cy.get("iframe[id=ketcher]");
-    cy.get("#compound-type-dropdown").select("ill-defined");
-    cy.get("iframe[id=marvin]");
-  });
-  it("marvinjs should have import/export functionality", () => {
-    cy.get("#compound-type-dropdown")
-      .select("ill-defined")
-      .then(() => {
-        cy.get("#marvin-export-button");
-        cy.get("#marvin-import-textarea");
-        cy.get("#marvin-import-button");
-      });
-  });
-  it("marvinjs should be able to import/export mrvfiles", () => {
-    var mrvfile =
-      '<cml xmlns="http://www.chemaxon.com" version="ChemAxon file format v20.9.0, generated by vunknown" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.chemaxon.com http://www.chemaxon.com/marvin/schema/mrvSchema_20_9_0.xsd"><MDocument><MChemicalStruct><molecule molID="m1"><atomArray><atom id="a1" elementType="O" x2="-0.3333333333333333" y2="0" lonePair="2"/></atomArray><bondArray/></molecule></MChemicalStruct><MElectronContainer occupation="0 0" radical="0" id="o1"><MElectron atomRefs="m1.a1" difLoc="0.0 0.0 0.0"/><MElectron atomRefs="m1.a1" difLoc="0.0 0.0 0.0"/></MElectronContainer><MElectronContainer occupation="0 0" radical="0" id="o2"><MElectron atomRefs="m1.a1" difLoc="0.0 0.0 0.0"/><MElectron atomRefs="m1.a1" difLoc="0.0 0.0 0.0"/></MElectronContainer></MDocument></cml>';
-    cy.get("#compound-type-dropdown")
-      .select("ill-defined")
-      .then(() => {
-        cy.get("#marvin-import-textarea")
-          .clear()
-          .invoke("val", mrvfile)
-          .trigger("input")
-          .then(() => {
-            cy.get("#marvin-import-button").click();
-            cy.get("#marvin-import-textarea")
-              .clear()
-              .trigger("input")
-              .then(() => {
-                cy.get("#marvin-export-button")
-                  .click()
-                  .then(() => {
-                    cy.get("#marvin-import-textarea")
-                      .invoke("val")
-                      .should("contain", mrvfile);
-                  });
-              });
-          });
-      });
-  });
-  it("should load defined compound into ketcher window", () => {
-    cy.get("[data-cy=search-box]").type("DTXCID302000003");
-    cy.get("[data-cy=search-button]").click();
-    cy.get("iframe[id=ketcher]");
-    cy.get("#ketcher-import-textarea")
-      .invoke("val")
-      .should("contain", "Ketcher")
-      .should("contain", "V2000");
-  });
-  it("should load illdefined compound into marvin window", () => {
-    cy.get("[data-cy=search-box]").type("DTXCID502000009");
-    cy.get("[data-cy=search-button]").click();
-    cy.get("iframe[id=marvin]");
-    cy.get("#marvin-import-textarea")
-      .invoke("val")
-      .should("contain", '<cml xmlns="http://www.chemaxon.com"')
-      .should("contain", "<MDocument><MChemicalStruct>");
-  });
-  it("should update the import textarea when ketcher changes", () => {
-    cy.get("iframe[id=ketcher]")
-      .its("0.contentDocument.body")
-      .should("not.be.empty");
-    cy.get("#ketcher-import-textarea")
-      .invoke("val")
-      .should("be.empty");
-    cy.get("iframe[id=ketcher]")
-      .its("0.contentDocument.body")
-      .should("not.be.empty")
-      .then(cy.wrap)
-      .find("#atom")
-      .find("button", "Hydrogen (H)")
-      .first()
-      .click();
-    cy.get("iframe[id=ketcher]")
-      .its("0.contentDocument.body")
-      .should("not.be.empty")
-      .then(cy.wrap)
-      .find("#canvas")
-      .click();
-    cy.get("#ketcher-import-textarea")
-      .invoke("val")
-      .should("not.be.empty");
-  });
-  it("bad search should alert invalidity", () => {
-    cy.get("[data-cy=search-box]").type("compound 47");
-    cy.get("[data-cy=search-button]").click();
-    cy.get("[data-cy=alert-box]").should("contain", "compound 47 not valid");
-  });
-  it("logout should provide message to user", () => {
-    cy.get("[data-cy=user-dropdown]").click();
-    cy.get("[data-cy=logout-button]").click();
-    cy.get("[data-cy=alert-box]").should(
-      "contain",
-      "karyn, you are no longer logged in!"
-    );
+  it("should display the search bar form to an authenticated user", () => {
+    cy.get('[data-cy=search-box]');
   });
 });
